Separated function file and test file
